refactor: save and retrieve monitors to state repo in dry-run, move MonitorBuilder to test-helpers
