+++ src/test_helpers/mod.rs
@@ -4,6 +4,7 @@ mod abi;
 mod block;
 mod http_client;
 mod log;
+mod monitor;
 mod monitor_manager;
 mod receipt;
 mod transaction;
@@ -12,8 +13,7 @@ pub use abi::{create_test_abi_service, erc20_abi_json};
 pub use block::BlockBuilder;
 pub use http_client::create_test_http_client;
 pub use log::LogBuilder;
+pub use monitor::MonitorBuilder;
 pub use monitor_manager::create_test_monitor_manager;
 pub use receipt::ReceiptBuilder;
 pub use transaction::TransactionBuilder;
-
-pub use crate::models::builder::MonitorBuilder;
